refactor scheduler tests.
