+++ tests/scheduler/test_dispatcher.py
@@ -15,15 +15,11 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-import os
 import unittest
 
 import cloudpickle
 from notification_service.event import EventKey, Event
 
-from ai_flow.common.util.db_util.db_migration import init_db
-from ai_flow.common.util.db_util.session import new_session
-from ai_flow.metadata.metadata_manager import MetadataManager
 from ai_flow.model.action import TaskAction
 from ai_flow.model.condition import Condition
 from ai_flow.model.execution_type import ExecutionType
@@ -34,19 +30,12 @@ from ai_flow.model.status import WorkflowStatus
 from ai_flow.model.workflow import Workflow
 from ai_flow.scheduler.dispatcher import Dispatcher
 from ai_flow.scheduler.worker import Worker
+from tests.scheduler.test_utils import UnitTestWithNamespace
 
 
-class TestDispatcher(unittest.TestCase):
+class TestDispatcher(UnitTestWithNamespace):
     def setUp(self) -> None:
-        self.file = 'test.db'
-        self._delete_db_file()
-        self.url = 'sqlite:///{}'.format(self.file)
-        init_db(self.url)
-        self.session = new_session(db_uri=self.url)
-        self.metadata_manager = MetadataManager(session=self.session)
-        self.namespace_name = 'namespace'
-        namespace_meta = self.metadata_manager.add_namespace(name=self.namespace_name, properties={'a': 'a'})
-        self.metadata_manager.flush()
+        super().setUp()
         with Workflow(name='workflow') as workflow:
             op = Operator(name='op')
             op.action_on_condition(action=TaskAction.START, condition=Condition(expect_events=[
@@ -86,14 +75,6 @@ class TestDispatcher(unittest.TestCase):
         self.metadata_manager.update_workflow_execution_status(workflow_execution_id=self.workflow_execution_meta.id,
                                                                status=WorkflowStatus.RUNNING)
         self.metadata_manager.flush()
-
-    def _delete_db_file(self):
-        if os.path.exists(self.file):
-            os.remove(self.file)
-
-    def tearDown(self) -> None:
-        self.session.close()
-        self._delete_db_file()
 
     def test_dispatch(self):
         worker_num = 3
