+++ tests/scheduler/test_rule_extractor.py
@@ -15,17 +15,12 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-import copy
 import json
-import os
 import unittest
 
 import cloudpickle
 from notification_service.event import Event, EventKey
 
-from ai_flow.common.util.db_util.db_migration import init_db
-from ai_flow.common.util.db_util.session import new_session
-from ai_flow.metadata.metadata_manager import MetadataManager
 from ai_flow.model.action import TaskAction
 from ai_flow.model.condition import Condition
 from ai_flow.model.execution_type import ExecutionType
@@ -36,6 +31,7 @@ from ai_flow.model.status import WorkflowStatus
 from ai_flow.model.workflow import Workflow
 from ai_flow.scheduler.rule_extractor import gen_all_combination, gen_all_tuple_by_event_key, \
     workflow_expect_event_tuples, build_task_rule_index, RuleExtractor
+from tests.scheduler.test_utils import UnitTestWithNamespace
 
 
 class TestRuleExtractorUtil(unittest.TestCase):
@@ -139,25 +135,7 @@ class TestRuleExtractorUtil(unittest.TestCase):
         return workflow_dict
 
 
-class TestRuleExtractor(unittest.TestCase):
-    def setUp(self) -> None:
-        self.file = 'test.db'
-        self._delete_db_file()
-        self.url = 'sqlite:///{}'.format(self.file)
-        init_db(self.url)
-        self.session = new_session(db_uri=self.url)
-        self.metadata_manager = MetadataManager(session=self.session)
-        self.namespace_name = 'namespace'
-        namespace_meta = self.metadata_manager.add_namespace(name=self.namespace_name, properties={'a': 'a'})
-        self.metadata_manager.flush()
-
-    def _delete_db_file(self):
-        if os.path.exists(self.file):
-            os.remove(self.file)
-
-    def tearDown(self) -> None:
-        self.session.close()
-        self._delete_db_file()
+class TestRuleExtractor(UnitTestWithNamespace):
 
     def test_extract_workflow_rules(self):
         def build_workflows():
